+++ bdtsim/cli/list_output_formats.py
@@ -15,14 +15,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .environment import Environment
-from .environment_manager import EnvironmentManager
-from .pyevm import PyEVMEnvironment
-from .web3_environments import Web3Environment
+import argparse
 
-__all__ = [
-    'Environment',
-    'EnvironmentManager',
-    'PyEVMEnvironment',
-    'Web3Environment'
-]
+from bdtsim.output import OutputFormatManager
+from .command_manager import SubCommand
+
+
+class ListOutputFormatsSubCommand(SubCommand):
+    def __call__(self, args: argparse.Namespace) -> None:
+        for name in OutputFormatManager.output_formats.keys():
+            print(name)