+++ tests/test_commands.py
@@ -45,6 +45,17 @@ class CommandTest(unittest.TestCase):
             'FileDataProvider'
         ]))
 
+    def test_list_output_formats(self):
+        p = subprocess.Popen(['bdtsim', 'list-output-formats'], stdout=subprocess.PIPE)
+        out, err = p.communicate()
+        self.assertEqual(p.returncode, 0)
+        self.assertEqual(out.decode('utf-8').strip(), '\n'.join([
+            'dot',
+            'human-readable',
+            'json',
+            'yaml'
+        ]))
+
     def test_run_simplepayment(self):
         p = subprocess.Popen(['bdtsim', 'run', 'SimplePayment', 'PyEVM'], stdout=subprocess.PIPE)
         out, err = p.communicate()