+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   * Added sub-command `environment-info`
   * Enabled DataProvider support
     * added `list-data-providers` command
+  * Improved output format support
+    * added `list-output-formats` command
 
 
 ## v0.1.2